excess files removed, naming convention fix
