+++ src/constants.ts
@@ -5,10 +5,9 @@ export let NIFTY_GATEWAY = Address.fromString("0x431bd1297a1c7664D599364a427A2d9
 export let RARIBLE = Address.fromString("0xd07dc4262BCDbf85190C01c996b4C06a461d2430");
 export let SUPERRARE = Address.fromString("0x41a322b28d0ff354040e2cbc676f0320d8c8850d");
 export let SUPERRARE_V2 = Address.fromString("0xb932a70A57673d89f4acfFBE830E8ed7f75Fb9e0");
-
 export let KNOW_ORIGIN = Address.fromString("0xFBeef911Dc5821886e1dda71586d90eD28174B7d");
-
 export let ZORA_NFT = Address.fromString("0xabEFBc9fD2F806065b4f3C237d4b59D9A97Bcac7");
+export let SORARE = Address.fromString("0x629a673a8242c2ac4b7b8c5d8735fbeac21a6205");
 
 export let ZERO_ADDRESS = Address.fromString("0x0000000000000000000000000000000000000000");
 export let BIGINT_ZERO = BigInt.fromI32(0);
