--- src/constants/dontbuymeme.ts
@@ -1,2 +1,0 @@
-import {  Address } from "@graphprotocol/graph-ts";
-export let DONTBUYMEME = Address.fromString("0xe4605d46Fd0B3f8329d936a8b258D69276cBa264");
