--- src/constants/sorare.ts
@@ -1,3 +1,0 @@
-import { Address } from "@graphprotocol/graph-ts";
-
-export let SORARE = Address.fromString("0x629a673a8242c2ac4b7b8c5d8735fbeac21a6205");
