+++ src/sorare.ts
@@ -1,17 +1,17 @@
-import { Transfer } from "../../generated/templates/NftContract/ERC721";
-import { Nft, NftContract } from "../../generated/schema";
+import { Transfer } from "../generated/templates/NftContract/ERC721";
+import { Nft, NftContract } from "../generated/schema";
+import { Sorare } from "../generated/Sorare/Sorare";
 import { fetchName, fetchSymbol, handleTransfer } from "./mapping";
-import { DontBuyMeme } from "../../generated/DontBuyMeme/DontBuyMeme";
-import { DONTBUYMEME, ZERO_ADDRESS } from "../constants";
 import { Address, BigInt, log } from "@graphprotocol/graph-ts";
+import { SORARE, ZERO_ADDRESS } from "./constants";
 
-export function handleTransferDontBuyMeme(event: Transfer): void {
+export function handleTransferSorare(event: Transfer): void {
   let address = event.address.toHexString();
   if (NftContract.load(address) == null) {
     let nftContract = new NftContract(address);
     nftContract.name = fetchName(event.address);
     nftContract.symbol = fetchSymbol(event.address);
-    nftContract.platform = "DontBuyMeme";
+    nftContract.platform = "Sorare";
     nftContract.save();
   }
 
@@ -26,11 +26,11 @@ export function handleTransferDontBuyMeme(event: Transfer): void {
 }
 
 function getAddressOfCreator(address: Address, tokenId: BigInt): Address {
-  if (address == DONTBUYMEME) {
-    let contract = DontBuyMeme.bind(address);
-    return contract.creators(tokenId);
+  if (address == SORARE) {
+    let contract = Sorare.bind(address);
+    return contract.ownerOf(tokenId);
   } else {
-    log.warning("OurZora contract address {} not found!", [address.toHexString()]);
+    log.warning("Sorare contract address {} not found!", [address.toHexString()]);
     return ZERO_ADDRESS;
   }
 }
